requirements.txt
torch
torchvision
numpy
pillow
matplotlib
wandb